refactor: removed rx_gps_data's params, removed rx_uart_data, set time/buf size
Removed rx_uart_data function is not needed since it is only used once and is
redundant to create another function to read UART data. Removed rx_gps_data's
parameters for now. May add parameters in the future to make the function more
customizable. Explicitly set the time and buffer value for the uart_read_bytes
function to fix variable type issues when providing values to the
uart_read_bytes function. Removed the UART_2 definition since it is
unnecessary.